+++ _test_yaml/test/ensure_build_test.dart
@@ -4,6 +4,8 @@
 
 @TestOn('vm')
 @Tags(['presubmit-only'])
+library;
+
 import 'package:build_verify/build_verify.dart';
 import 'package:test/test.dart';