+++ json_serializable/lib/src/decode_helper.dart
@@ -21,7 +21,7 @@ class CreateFactoryResult {
   CreateFactoryResult(this.output, this.usedFields);
 }
 
-abstract class DecodeHelper implements HelperCore {
+mixin DecodeHelper implements HelperCore {
   CreateFactoryResult createFactory(
     Map<String, FieldElement> accessibleFields,
     Map<String, String> unavailableReasons,
@@ -31,7 +31,7 @@ abstract class DecodeHelper implements HelperCore {
 
     final mapType = config.anyMap ? 'Map' : 'Map<String, dynamic>';
     buffer.write('$targetClassReference '
-        '${prefix}FromJson${genericClassArgumentsImpl(true)}'
+        '${prefix}FromJson${genericClassArgumentsImpl(withConstraints: true)}'
         '($mapType json');
 
     if (config.genericArgumentFactories) {
@@ -382,27 +382,18 @@ _ConstructorData _writeConstructorInvocation(
       '$className'
       '${genericClassArguments(classElement, false)}'
       '$constructorExtra(',
-    );
-  if (constructorArguments.isNotEmpty) {
-    buffer
-      ..writeln()
-      ..writeAll(constructorArguments.map((paramElement) {
-        final content =
-            deserializeForField(paramElement.name, ctorParam: paramElement);
-        return '      $content,\n';
-      }));
-  }
-  if (namedConstructorArguments.isNotEmpty) {
-    buffer
-      ..writeln()
-      ..writeAll(namedConstructorArguments.map((paramElement) {
-        final value =
-            deserializeForField(paramElement.name, ctorParam: paramElement);
-        return '      ${paramElement.name}: $value,\n';
-      }));
-  }
-
-  buffer.write(')');
+    )
+    ..writeAll(constructorArguments.map((paramElement) {
+      final content =
+          deserializeForField(paramElement.name, ctorParam: paramElement);
+      return '      $content,\n';
+    }))
+    ..writeAll(namedConstructorArguments.map((paramElement) {
+      final value =
+          deserializeForField(paramElement.name, ctorParam: paramElement);
+      return '      ${paramElement.name}: $value,\n';
+    }))
+    ..write(')');
 
   usedCtorParamsAndFields.addAll(remainingFieldsForInvocationBody);