+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -11,9 +11,11 @@ class KeyConfig {
 
   final bool disallowNullValue;
 
-  final bool ignore;
+  final bool? includeFromJson;
 
   final bool includeIfNull;
+
+  final bool? includeToJson;
 
   final String name;
 
@@ -26,8 +28,9 @@ class KeyConfig {
   KeyConfig({
     required this.defaultValue,
     required this.disallowNullValue,
-    required this.ignore,
+    required this.includeFromJson,
     required this.includeIfNull,
+    required this.includeToJson,
     required this.name,
     required this.readValueFunctionName,
     required this.required,
@@ -46,6 +49,8 @@ class ClassConfig {
   final bool createFactory;
   final bool createToJson;
   final bool createFieldMap;
+  final bool createJsonKeys;
+  final bool createPerFieldToJson;
   final bool disallowUnrecognizedKeys;
   final bool explicitToJson;
   final FieldRename fieldRename;
@@ -63,6 +68,8 @@ class ClassConfig {
     required this.createFactory,
     required this.createToJson,
     required this.createFieldMap,
+    required this.createJsonKeys,
+    required this.createPerFieldToJson,
     required this.disallowUnrecognizedKeys,
     required this.explicitToJson,
     required this.fieldRename,
@@ -82,6 +89,10 @@ class ClassConfig {
         constructor: config.constructor ?? ClassConfig.defaults.constructor,
         createFieldMap:
             config.createFieldMap ?? ClassConfig.defaults.createFieldMap,
+        createJsonKeys:
+            config.createJsonKeys ?? ClassConfig.defaults.createJsonKeys,
+        createPerFieldToJson: config.createPerFieldToJson ??
+            ClassConfig.defaults.createPerFieldToJson,
         createFactory:
             config.createFactory ?? ClassConfig.defaults.createFactory,
         createToJson: config.createToJson ?? ClassConfig.defaults.createToJson,
@@ -110,6 +121,8 @@ class ClassConfig {
     createFactory: true,
     createToJson: true,
     createFieldMap: false,
+    createJsonKeys: false,
+    createPerFieldToJson: false,
     disallowUnrecognizedKeys: false,
     explicitToJson: false,
     fieldRename: FieldRename.none,
@@ -120,19 +133,21 @@ class ClassConfig {
   );
 
   JsonSerializable toJsonSerializable() => JsonSerializable(
-      checked: checked,
-      anyMap: anyMap,
-      constructor: constructor,
-      createFactory: createFactory,
-      createToJson: createToJson,
-      createFieldMap: createFieldMap,
-      ignoreUnannotated: ignoreUnannotated,
-      explicitToJson: explicitToJson,
-      includeIfNull: includeIfNull,
-      genericArgumentFactories: genericArgumentFactories,
-      fieldRename: fieldRename,
-      disallowUnrecognizedKeys: disallowUnrecognizedKeys,
-      defaultOnException: defaultOnException,
-      // TODO typeConverters = []
+        checked: checked,
+        anyMap: anyMap,
+        constructor: constructor,
+        createFactory: createFactory,
+        createToJson: createToJson,
+        createFieldMap: createFieldMap,
+        createJsonKeys: createJsonKeys,
+        createPerFieldToJson: createPerFieldToJson,
+        ignoreUnannotated: ignoreUnannotated,
+        explicitToJson: explicitToJson,
+        includeIfNull: includeIfNull,
+        genericArgumentFactories: genericArgumentFactories,
+        fieldRename: fieldRename,
+        disallowUnrecognizedKeys: disallowUnrecognizedKeys,
+        defaultOnException: defaultOnException,
+        // TODO typeConverters = []
       );
 }