+++ json_serializable/test/kitchen_sink/kitchen_sink.g_any_map.dart
@@ -184,6 +184,8 @@ class KitchenSink implements k.KitchenSink {
     _validatedPropertyNo42 = value;
   }
 
+  k.RecordSample? recordField;
+
   bool operator ==(Object other) => k.sinkEquals(this, other);
 
   static Object? _trickyValueAccessor(Map json, String key) {
@@ -249,6 +251,7 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
 @JsonSerializable(
   anyMap: true,
 )
+// ignore: inference_failure_on_instance_creation
 @GenericConverter()
 class JsonConverterGeneric<S, T, U> {
   S item;