+++ json_serializable/test/kitchen_sink/kitchen_sink.g_exclude_null.dart
@@ -186,6 +186,8 @@ class KitchenSink implements k.KitchenSink {
     _validatedPropertyNo42 = value;
   }
 
+  k.RecordSample? recordField;
+
   bool operator ==(Object other) => k.sinkEquals(this, other);
 
   static Object? _trickyValueAccessor(Map json, String key) {
@@ -251,6 +253,7 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
 @JsonSerializable(
   includeIfNull: false,
 )
+// ignore: inference_failure_on_instance_creation
 @GenericConverter()
 class JsonConverterGeneric<S, T, U> {
   S item;