+++ json_serializable/test/src/to_from_json_test_input.dart
@@ -2,6 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart=3.6
+
+// ignore_for_file: strict_top_level_inference
+
 part of '_json_serializable_test_input.dart';
 
 int _toInt(bool input) => 42;
@@ -122,7 +126,7 @@ class Reproduce869NullableGenericTypeWithDefault {
 }
 
 List<int>? _fromList(List? pairs) =>
-    pairs?.map((it) => it as int).toList(growable: false);
+    pairs?.map((it) => (it as num).toInt()).toList(growable: false);
 
 List<List>? _toList(List<int>? pairs) =>
     pairs?.map((it) => [it]).toList(growable: false);
@@ -175,18 +179,11 @@ String? _toStringNullOnEmpty(String input) => input.isEmpty ? null : input;
 @ShouldGenerate(
   r'''
 Map<String, dynamic> _$ToJsonNullableFalseIncludeIfNullFalseToJson(
-    ToJsonNullableFalseIncludeIfNullFalse instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('field', _toStringNullOnEmpty(instance.field));
-  return val;
-}
+        ToJsonNullableFalseIncludeIfNullFalse instance) =>
+    <String, dynamic>{
+      if (_toStringNullOnEmpty(instance.field) case final value?)
+        'field': value,
+    };
 ''',
 )
 @JsonSerializable(createFactory: false)