+++ json_serializable/test/supported_types/input.type_list.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-// ignore_for_file: lines_longer_than_80_chars, text_direction_code_point_in_literal
+// ignore_for_file: lines_longer_than_80_chars, text_direction_code_point_in_literal, inference_failure_on_function_invocation, inference_failure_on_collection_literal
 
 part of 'input.type_list.dart';
 
@@ -249,7 +249,7 @@ SimpleClassOfDuration _$SimpleClassOfDurationFromJson(
         Map<String, dynamic> json) =>
     SimpleClassOfDuration(
       (json['value'] as List<dynamic>)
-          .map((e) => Duration(microseconds: e as int))
+          .map((e) => Duration(microseconds: (e as num).toInt()))
           .toList(),
     );
 
@@ -263,7 +263,7 @@ SimpleClassNullableOfDuration _$SimpleClassNullableOfDurationFromJson(
         Map<String, dynamic> json) =>
     SimpleClassNullableOfDuration(
       (json['value'] as List<dynamic>?)
-          ?.map((e) => Duration(microseconds: e as int))
+          ?.map((e) => Duration(microseconds: (e as num).toInt()))
           .toList(),
     );
 
@@ -277,7 +277,8 @@ SimpleClassOfDurationNullable _$SimpleClassOfDurationNullableFromJson(
         Map<String, dynamic> json) =>
     SimpleClassOfDurationNullable(
       (json['value'] as List<dynamic>)
-          .map((e) => e == null ? null : Duration(microseconds: e as int))
+          .map((e) =>
+              e == null ? null : Duration(microseconds: (e as num).toInt()))
           .toList(),
     );
 
@@ -292,7 +293,8 @@ SimpleClassNullableOfDurationNullable
             Map<String, dynamic> json) =>
         SimpleClassNullableOfDurationNullable(
           (json['value'] as List<dynamic>?)
-              ?.map((e) => e == null ? null : Duration(microseconds: e as int))
+              ?.map((e) =>
+                  e == null ? null : Duration(microseconds: (e as num).toInt()))
               .toList(),
         );
 
@@ -390,9 +392,97 @@ Map<String, dynamic> _$SimpleClassNullableOfEnumTypeNullableToJson(
       'value': instance.value?.map((e) => _$EnumTypeEnumMap[e]).toList(),
     };
 
+SimpleClassOfFromJsonDynamicParam _$SimpleClassOfFromJsonDynamicParamFromJson(
+        Map<String, dynamic> json) =>
+    SimpleClassOfFromJsonDynamicParam(
+      (json['value'] as List<dynamic>)
+          .map(FromJsonDynamicParam.fromJson)
+          .toList(),
+    );
+
+Map<String, dynamic> _$SimpleClassOfFromJsonDynamicParamToJson(
+        SimpleClassOfFromJsonDynamicParam instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };
+
+SimpleClassNullableOfFromJsonDynamicParam
+    _$SimpleClassNullableOfFromJsonDynamicParamFromJson(
+            Map<String, dynamic> json) =>
+        SimpleClassNullableOfFromJsonDynamicParam(
+          (json['value'] as List<dynamic>?)
+              ?.map(FromJsonDynamicParam.fromJson)
+              .toList(),
+        );
+
+Map<String, dynamic> _$SimpleClassNullableOfFromJsonDynamicParamToJson(
+        SimpleClassNullableOfFromJsonDynamicParam instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };
+
+SimpleClassOfFromJsonNullableObjectParam
+    _$SimpleClassOfFromJsonNullableObjectParamFromJson(
+            Map<String, dynamic> json) =>
+        SimpleClassOfFromJsonNullableObjectParam(
+          (json['value'] as List<dynamic>)
+              .map(FromJsonNullableObjectParam.fromJson)
+              .toList(),
+        );
+
+Map<String, dynamic> _$SimpleClassOfFromJsonNullableObjectParamToJson(
+        SimpleClassOfFromJsonNullableObjectParam instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };
+
+SimpleClassNullableOfFromJsonNullableObjectParam
+    _$SimpleClassNullableOfFromJsonNullableObjectParamFromJson(
+            Map<String, dynamic> json) =>
+        SimpleClassNullableOfFromJsonNullableObjectParam(
+          (json['value'] as List<dynamic>?)
+              ?.map(FromJsonNullableObjectParam.fromJson)
+              .toList(),
+        );
+
+Map<String, dynamic> _$SimpleClassNullableOfFromJsonNullableObjectParamToJson(
+        SimpleClassNullableOfFromJsonNullableObjectParam instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };
+
+SimpleClassOfFromJsonObjectParam _$SimpleClassOfFromJsonObjectParamFromJson(
+        Map<String, dynamic> json) =>
+    SimpleClassOfFromJsonObjectParam(
+      (json['value'] as List<dynamic>)
+          .map((e) => FromJsonObjectParam.fromJson(e as Object))
+          .toList(),
+    );
+
+Map<String, dynamic> _$SimpleClassOfFromJsonObjectParamToJson(
+        SimpleClassOfFromJsonObjectParam instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };
+
+SimpleClassNullableOfFromJsonObjectParam
+    _$SimpleClassNullableOfFromJsonObjectParamFromJson(
+            Map<String, dynamic> json) =>
+        SimpleClassNullableOfFromJsonObjectParam(
+          (json['value'] as List<dynamic>?)
+              ?.map((e) => FromJsonObjectParam.fromJson(e as Object))
+              .toList(),
+        );
+
+Map<String, dynamic> _$SimpleClassNullableOfFromJsonObjectParamToJson(
+        SimpleClassNullableOfFromJsonObjectParam instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };
+
 SimpleClassOfInt _$SimpleClassOfIntFromJson(Map<String, dynamic> json) =>
     SimpleClassOfInt(
-      (json['value'] as List<dynamic>).map((e) => e as int).toList(),
+      (json['value'] as List<dynamic>).map((e) => (e as num).toInt()).toList(),
     );
 
 Map<String, dynamic> _$SimpleClassOfIntToJson(SimpleClassOfInt instance) =>
@@ -403,7 +493,9 @@ Map<String, dynamic> _$SimpleClassOfIntToJson(SimpleClassOfInt instance) =>
 SimpleClassNullableOfInt _$SimpleClassNullableOfIntFromJson(
         Map<String, dynamic> json) =>
     SimpleClassNullableOfInt(
-      (json['value'] as List<dynamic>?)?.map((e) => e as int).toList(),
+      (json['value'] as List<dynamic>?)
+          ?.map((e) => (e as num).toInt())
+          .toList(),
     );
 
 Map<String, dynamic> _$SimpleClassNullableOfIntToJson(
@@ -415,7 +507,9 @@ Map<String, dynamic> _$SimpleClassNullableOfIntToJson(
 SimpleClassOfIntNullable _$SimpleClassOfIntNullableFromJson(
         Map<String, dynamic> json) =>
     SimpleClassOfIntNullable(
-      (json['value'] as List<dynamic>).map((e) => e as int?).toList(),
+      (json['value'] as List<dynamic>)
+          .map((e) => (e as num?)?.toInt())
+          .toList(),
     );
 
 Map<String, dynamic> _$SimpleClassOfIntNullableToJson(
@@ -427,7 +521,9 @@ Map<String, dynamic> _$SimpleClassOfIntNullableToJson(
 SimpleClassNullableOfIntNullable _$SimpleClassNullableOfIntNullableFromJson(
         Map<String, dynamic> json) =>
     SimpleClassNullableOfIntNullable(
-      (json['value'] as List<dynamic>?)?.map((e) => e as int?).toList(),
+      (json['value'] as List<dynamic>?)
+          ?.map((e) => (e as num?)?.toInt())
+          .toList(),
     );
 
 Map<String, dynamic> _$SimpleClassNullableOfIntNullableToJson(
@@ -527,6 +623,65 @@ Map<String, dynamic> _$SimpleClassNullableOfObjectNullableToJson(
         SimpleClassNullableOfObjectNullable instance) =>
     <String, dynamic>{
       'value': instance.value,
+    };
+
+SimpleClassOfRecord _$SimpleClassOfRecordFromJson(Map<String, dynamic> json) =>
+    SimpleClassOfRecord(
+      (json['value'] as List<dynamic>)
+          .map((e) => _$recordConvert(
+                e,
+                ($jsonValue) => (
+                  ($jsonValue[r'$1'] as num).toInt(),
+                  $jsonValue[r'$2'] as String,
+                  truth: $jsonValue['truth'] as bool,
+                ),
+              ))
+          .toList(),
+    );
+
+Map<String, dynamic> _$SimpleClassOfRecordToJson(
+        SimpleClassOfRecord instance) =>
+    <String, dynamic>{
+      'value': instance.value
+          .map((e) => <String, dynamic>{
+                r'$1': e.$1,
+                r'$2': e.$2,
+                'truth': e.truth,
+              })
+          .toList(),
+    };
+
+$Rec _$recordConvert<$Rec>(
+  Object? value,
+  $Rec Function(Map) convert,
+) =>
+    convert(value as Map<String, dynamic>);
+
+SimpleClassNullableOfRecord _$SimpleClassNullableOfRecordFromJson(
+        Map<String, dynamic> json) =>
+    SimpleClassNullableOfRecord(
+      (json['value'] as List<dynamic>?)
+          ?.map((e) => _$recordConvert(
+                e,
+                ($jsonValue) => (
+                  ($jsonValue[r'$1'] as num).toInt(),
+                  $jsonValue[r'$2'] as String,
+                  truth: $jsonValue['truth'] as bool,
+                ),
+              ))
+          .toList(),
+    );
+
+Map<String, dynamic> _$SimpleClassNullableOfRecordToJson(
+        SimpleClassNullableOfRecord instance) =>
+    <String, dynamic>{
+      'value': instance.value
+          ?.map((e) => <String, dynamic>{
+                r'$1': e.$1,
+                r'$2': e.$2,
+                'truth': e.truth,
+              })
+          .toList(),
     };
 
 SimpleClassOfString _$SimpleClassOfStringFromJson(Map<String, dynamic> json) =>