+++ json_serializable/tool/test_builder.dart
@@ -6,26 +6,25 @@ import 'dart:async';
 import 'dart:collection';
 
 import 'package:build/build.dart';
-import 'package:dart_style/dart_style.dart';
 import 'package:path/path.dart' as p;
 
 import 'shared.dart';
 
-final _formatter = DartFormatter();
-
-Builder testBuilder([_]) => validate('_test_builder', const _TestBuilder());
+Builder testBuilder([BuilderOptions? _]) =>
+    validate('_test_builder', const _TestBuilder());
 
 class _TestBuilder implements Builder {
   const _TestBuilder();
 
   @override
   FutureOr<void> build(BuildStep buildStep) async {
+    final formatter = await buildStep.formatter();
+
     final baseName = p.basenameWithoutExtension(buildStep.inputId.path);
 
     final sourceContent = await buildStep.readAsString(buildStep.inputId);
 
-    final factories =
-        SplayTreeMap.from({'$_kitchenSinkBaseName.dart': 'normal'});
+    final factories = SplayTreeMap.of({'$_kitchenSinkBaseName.dart': 'normal'});
 
     for (var config in _fileConfigurationMap[baseName]!) {
       final extension = _configToExtension(config);
@@ -56,7 +55,7 @@ class _TestBuilder implements Builder {
 
       final content = Replacement.generate(sourceContent, replacements);
 
-      await buildStep.writeAsString(newId, _formatter.format(content));
+      await buildStep.writeAsString(newId, formatter.format(content));
     }
 
     if (baseName == _kitchenSinkBaseName) {
@@ -73,7 +72,7 @@ class _TestBuilder implements Builder {
         '];',
       ];
 
-      await buildStep.writeAsString(newId, _formatter.format(lines.join('\n')));
+      await buildStep.writeAsString(newId, formatter.format(lines.join('\n')));
     }
   }