+++ json_annotation/CHANGELOG.md
@@ -1,7 +1,28 @@
-## 4.7.1-dev
+## 4.9.1-wip
 
+- Require Dart 3.6
+
+## 4.9.0
+
+- Require Dart 3.0
+- Add `JsonSerializable(createJsonKeys: true)`.
+  ([#1401](https://github.com/google/json_serializable.dart/pull/1401))
+
+## 4.8.1
+
+- Require Dart 2.19
+- Add topics.
+
+## 4.8.0
+
+- DEPRECATED `JsonKey.ignore`. Replaced by...
+- Added `JsonKey.includeFromJson` and `JsonKey.includeToJson` to allow 
+  fine-grained control of if a field is encoded/decoded.  
+- Added `JsonSerializable.createPerFieldToJson` which allows generating
+  a `_$ModelPerFieldToJson`, enabling partial encoding of a model.
 - Update `JsonKey` documentation to align with new features in 
   `package:json_serializable`.
+- Require Dart SDK `>=2.18.0`.
 
 ## 4.7.0